requirements.txt
mysql-connector-python==9.2.0
tk==0.1.0
python-dotenv

